Etch A Sketch
Create a program that simulates an etch a sketch machine on the gfxhat.
Use the function getch from the click module to capture a keystroke
The key s is use to start again a new drawing on the hat.
The key q is used to quit.
The character code for the arrow keys are the following:
'\x1b[A' UP arrow key
'\x1b[B' DOWN arrow key
'\x1b[C' RIGHT arrow key
'\x1b[D' LEFT arro key